// src/scenes/play_scene/pipe_manager.d.ts
export declare class PipesManager {
    static placePipe(uPipe: any, lPipe: any, difficulty: any, pipes: any, pipes_speed: number): void;
    static getRightMostPipe(pipes: any): number;
    static createPipes(pipes: any, dificulty: any, velocity: number): void;
}
//# sourceMappingURL=pipe_manager.d.ts.map